fix: 修复pb.go问题&增加请求延时

+++ external_api/baseinterface/internal/logic/BaseInterface/userloginlogic.go
@@ -52,7 +52,7 @@ func (l *UserloginLogic) Userlogin(req *types.UserloginHandlerRequest) (resp *ty
 	return &types.UserloginHandlerResponse{
 		StatusCode: 0,
 		StatusMsg:  "登录成功",
-		UserID:     uid.UserId,
+		UserID:     int64(uid.UserId),
 		Token:      TokenString,
 	}, err
 }
